requirements.txt
datasets
pandas
pyarrow
python-dotenv
groq
fastapi
uvicorn
pytest
